Fix date + checkbox schemas in user editing forms
In addition to updating the user editing forms in the admin route and in
profile settings to accept empty values from the date fields, this also
modifies the database action so that in these cases, it will explicitly
null the birthdate fields for the relevant records. TODO: I realized
that I ought to reconfigure the optional number fields so that they can
be explicitly nullable in the same way.